renaming old input
