+++ projects/dxc-ngx-cdk/src/lib/dxc-input-text/dxc-input-prefix-icon/dxc-input-prefix-icon.component.html
