+++ projects/dxc-ngx-cdk/src/lib/dxc-input-text/dxc-input-suffix-icon/dxc-input-suffix-icon.component.ts
