+++ projects/dxc-ngx-cdk/src/lib/dxc-slider/dxc-slider.module.ts
@@ -3,7 +3,7 @@ import { MatSliderModule } from "@angular/material/slider";
 import { FormsModule } from "@angular/forms";
 import { DxcSliderComponent } from "./dxc-slider.component";
 import { CommonModule } from "@angular/common";
-import { DxcInputTextModule } from "../dxc-text-input/dxc-input-text.module";
+import { DxcInputTextModule } from "../dxc-input-text/dxc-input-text.module";
 
 @NgModule({
   declarations: [DxcSliderComponent],
